feat: User DDD 시작 (regex 오류있음)

+++ src/main/kotlin/mashup/spring/seehyang/controller/api/dto/user/RegisterUserDetailDto.kt
@@ -4,7 +4,7 @@ import mashup.spring.seehyang.domain.entity.perfume.Gender
 
 data class RegisterUserDetailRequest(
     val gender: Gender,
-    val age: Short,
+    val age: Int,
     val nickname: String,
 )
